Move KafkaClient into core-backup
