+++ core-backup/src/main/scala/io/aiven/guardian/kafka/backup/KafkaClient.scala
@@ -1,4 +1,4 @@
-package io.aiven.guardian.kafka
+package io.aiven.guardian.kafka.backup
 
 import akka.Done
 import akka.actor.ActorSystem
