+++ core-backup/src/test/scala/io/aiven/guardian/kafka/backup/MockedKafkaClientInterface.scala
@@ -1,4 +1,4 @@
-package io.aiven.guardian.kafka
+package io.aiven.guardian.kafka.backup
 
 import akka.Done
 import akka.NotUsed
